Move page-header and search-result-pagination into commons
