+++ frontend/typescript/components/commons/search-result-pagination/search-result-pagination.component.html
